Minor override file reorg
